change math routines from int to size_t to allow big size input also: tsdf data layer allow classification switch

+++ include/caffe/blob.hpp
@@ -73,7 +73,7 @@ class Blob {
     return shape_[CanonicalAxisIndex(index)];
   }
   inline int num_axes() const { return shape_.size(); }
-  inline int count() const { return count_; }
+  inline size_t count() const { return count_; }
 
   /**
    * @brief Compute the volume of a slice; i.e., the product of dimensions
@@ -290,8 +290,8 @@ class Blob {
   shared_ptr<SyncedMemory> acum_diff_;
   shared_ptr<SyncedMemory> shape_data_;
   vector<int> shape_;
-  int count_;
-  int capacity_;
+  size_t count_;
+  size_t capacity_;
 
   DISABLE_COPY_AND_ASSIGN(Blob);
 };  // class Blob
