+++ include/caffe/util/math_functions.hpp
@@ -26,46 +26,46 @@ void caffe_cpu_gemv(const CBLAS_TRANSPOSE TransA, const int M, const int N,
     Dtype* y);
 
 template <typename Dtype>
-void caffe_axpy(const int N, const Dtype alpha, const Dtype* X,
+void caffe_axpy(const size_t N, const Dtype alpha, const Dtype* X,
     Dtype* Y);
 
 template <typename Dtype>
-void caffe_cpu_axpby(const int N, const Dtype alpha, const Dtype* X,
+void caffe_cpu_axpby(const size_t N, const Dtype alpha, const Dtype* X,
     const Dtype beta, Dtype* Y);
 
 template <typename Dtype>
-void caffe_copy(const int N, const Dtype *X, Dtype *Y);
+void caffe_copy(const size_t N, const Dtype *X, Dtype *Y);
 
 template <typename Dtype>
-void caffe_set(const int N, const Dtype alpha, Dtype *X);
+void caffe_set(const size_t N, const Dtype alpha, Dtype *X);
 
 inline void caffe_memset(const size_t N, const int alpha, void* X) {
   memset(X, alpha, N);  // NOLINT(caffe/alt_fn)
 }
 
 template <typename Dtype>
-void caffe_add_scalar(const int N, const Dtype alpha, Dtype *X);
+void caffe_add_scalar(const size_t N, const Dtype alpha, Dtype *X);
 
 template <typename Dtype>
-void caffe_scal(const int N, const Dtype alpha, Dtype *X);
+void caffe_scal(const size_t N, const Dtype alpha, Dtype *X);
 
 template <typename Dtype>
-void caffe_sqr(const int N, const Dtype* a, Dtype* y);
+void caffe_sqr(const size_t N, const Dtype* a, Dtype* y);
 
 template <typename Dtype>
-void caffe_add(const int N, const Dtype* a, const Dtype* b, Dtype* y);
+void caffe_add(const size_t N, const Dtype* a, const Dtype* b, Dtype* y);
 
 template <typename Dtype>
-void caffe_sub(const int N, const Dtype* a, const Dtype* b, Dtype* y);
+void caffe_sub(const size_t N, const Dtype* a, const Dtype* b, Dtype* y);
 
 template <typename Dtype>
-void caffe_mul(const int N, const Dtype* a, const Dtype* b, Dtype* y);
+void caffe_mul(const size_t N, const Dtype* a, const Dtype* b, Dtype* y);
 
 template <typename Dtype>
-void caffe_div(const int N, const Dtype* a, const Dtype* b, Dtype* y);
+void caffe_div(const size_t N, const Dtype* a, const Dtype* b, Dtype* y);
 
 template <typename Dtype>
-void caffe_powx(const int n, const Dtype* a, const Dtype b, Dtype* y);
+void caffe_powx(const size_t n, const Dtype* a, const Dtype b, Dtype* y);
 
 unsigned int caffe_rng_rand();
 
@@ -73,37 +73,37 @@ template <typename Dtype>
 Dtype caffe_nextafter(const Dtype b);
 
 template <typename Dtype>
-void caffe_rng_uniform(const int n, const Dtype a, const Dtype b, Dtype* r);
+void caffe_rng_uniform(const size_t n, const Dtype a, const Dtype b, Dtype* r);
 
 template <typename Dtype>
-void caffe_rng_gaussian(const int n, const Dtype mu, const Dtype sigma,
+void caffe_rng_gaussian(const size_t n, const Dtype mu, const Dtype sigma,
                         Dtype* r);
 
 template <typename Dtype>
-void caffe_rng_bernoulli(const int n, const Dtype p, int* r);
+void caffe_rng_bernoulli(const size_t n, const Dtype p, int* r);
 
 template <typename Dtype>
-void caffe_rng_bernoulli(const int n, const Dtype p, unsigned int* r);
+void caffe_rng_bernoulli(const size_t n, const Dtype p, unsigned int* r);
 
 template <typename Dtype>
-void caffe_exp(const int n, const Dtype* a, Dtype* y);
+void caffe_exp(const size_t n, const Dtype* a, Dtype* y);
 
 template <typename Dtype>
-void caffe_abs(const int n, const Dtype* a, Dtype* y);
+void caffe_abs(const size_t n, const Dtype* a, Dtype* y);
 
 template <typename Dtype>
-Dtype caffe_cpu_dot(const int n, const Dtype* x, const Dtype* y);
+Dtype caffe_cpu_dot(const size_t n, const Dtype* x, const Dtype* y);
 
 template <typename Dtype>
-Dtype caffe_cpu_strided_dot(const int n, const Dtype* x, const int incx,
+Dtype caffe_cpu_strided_dot(const size_t n, const Dtype* x, const int incx,
     const Dtype* y, const int incy);
 
 template <typename Dtype>
-int caffe_cpu_hamming_distance(const int n, const Dtype* x, const Dtype* y);
+int caffe_cpu_hamming_distance(const size_t n, const Dtype* x, const Dtype* y);
 
 // Returns the sum of the absolute values of the elements of vector x
 template <typename Dtype>
-Dtype caffe_cpu_asum(const int n, const Dtype* x);
+Dtype caffe_cpu_asum(const size_t n, const Dtype* x);
 
 // the branchless, type-safe version from
 // http://stackoverflow.com/questions/1903954/is-there-a-standard-sign-function-signum-sgn-in-c-c
@@ -120,7 +120,7 @@ inline int8_t caffe_sign(Dtype val) {
 // So they have to be pasted here temporarily.
 #define DEFINE_CAFFE_CPU_UNARY_FUNC(name, operation) \
   template<typename Dtype> \
-  void caffe_cpu_##name(const int n, const Dtype* x, Dtype* y) { \
+  void caffe_cpu_##name(const size_t n, const Dtype* x, Dtype* y) { \
     CHECK_GT(n, 0); CHECK(x); CHECK(y); \
     for (int i = 0; i < n; ++i) { \
       operation; \
@@ -140,7 +140,7 @@ DEFINE_CAFFE_CPU_UNARY_FUNC(sgnbit, \
 DEFINE_CAFFE_CPU_UNARY_FUNC(fabs, y[i] = std::fabs(x[i]));
 
 template <typename Dtype>
-void caffe_cpu_scale(const int n, const Dtype alpha, const Dtype *x, Dtype* y);
+void caffe_cpu_scale(const size_t n, const Dtype alpha, const Dtype *x, Dtype* y);
 
 #ifndef CPU_ONLY  // GPU
 
@@ -159,17 +159,17 @@ void caffe_gpu_gemv(const CBLAS_TRANSPOSE TransA, const int M, const int N,
     Dtype* y);
 
 template <typename Dtype>
-void caffe_gpu_axpy(const int N, const Dtype alpha, const Dtype* X,
+void caffe_gpu_axpy(const size_t N, const Dtype alpha, const Dtype* X,
     Dtype* Y);
 
 template <typename Dtype>
-void caffe_gpu_axpby(const int N, const Dtype alpha, const Dtype* X,
+void caffe_gpu_axpby(const size_t N, const Dtype alpha, const Dtype* X,
     const Dtype beta, Dtype* Y);
 
 void caffe_gpu_memcpy(const size_t N, const void *X, void *Y);
 
 template <typename Dtype>
-void caffe_gpu_set(const int N, const Dtype alpha, Dtype *X);
+void caffe_gpu_set(const size_t N, const Dtype alpha, Dtype *X);
 
 inline void caffe_gpu_memset(const size_t N, const int alpha, void* X) {
 #ifndef CPU_ONLY
@@ -180,35 +180,35 @@ inline void caffe_gpu_memset(const size_t N, const int alpha, void* X) {
 }
 
 template <typename Dtype>
-void caffe_gpu_add_scalar(const int N, const Dtype alpha, Dtype *X);
+void caffe_gpu_add_scalar(const size_t N, const Dtype alpha, Dtype *X);
 
 template <typename Dtype>
-void caffe_gpu_scal(const int N, const Dtype alpha, Dtype *X);
+void caffe_gpu_scal(const size_t N, const Dtype alpha, Dtype *X);
 
 template <typename Dtype>
-void caffe_gpu_add(const int N, const Dtype* a, const Dtype* b, Dtype* y);
+void caffe_gpu_add(const size_t N, const Dtype* a, const Dtype* b, Dtype* y);
 
 template <typename Dtype>
-void caffe_gpu_sub(const int N, const Dtype* a, const Dtype* b, Dtype* y);
+void caffe_gpu_sub(const size_t N, const Dtype* a, const Dtype* b, Dtype* y);
 
 template <typename Dtype>
-void caffe_gpu_mul(const int N, const Dtype* a, const Dtype* b, Dtype* y);
+void caffe_gpu_mul(const size_t N, const Dtype* a, const Dtype* b, Dtype* y);
 
 template <typename Dtype>
-void caffe_gpu_div(const int N, const Dtype* a, const Dtype* b, Dtype* y);
+void caffe_gpu_div(const size_t N, const Dtype* a, const Dtype* b, Dtype* y);
 
 template <typename Dtype>
-void caffe_gpu_abs(const int n, const Dtype* a, Dtype* y);
+void caffe_gpu_abs(const size_t n, const Dtype* a, Dtype* y);
 
 template <typename Dtype>
-void caffe_gpu_exp(const int n, const Dtype* a, Dtype* y);
+void caffe_gpu_exp(const size_t n, const Dtype* a, Dtype* y);
 
 template <typename Dtype>
-void caffe_gpu_powx(const int n, const Dtype* a, const Dtype b, Dtype* y);
+void caffe_gpu_powx(const size_t n, const Dtype* a, const Dtype b, Dtype* y);
 
 // caffe_gpu_rng_uniform with two arguments generates integers in the range
 // [0, UINT_MAX].
-void caffe_gpu_rng_uniform(const int n, unsigned int* r);
+void caffe_gpu_rng_uniform(const size_t n, unsigned int* r);
 
 // caffe_gpu_rng_uniform with four arguments generates floats in the range
 // (a, b] (strictly greater than a, less than or equal to b) due to the
@@ -216,52 +216,52 @@ void caffe_gpu_rng_uniform(const int n, unsigned int* r);
 // curandGenerateUniform; with other limits will shift and scale the outputs
 // appropriately after calling curandGenerateUniform.
 template <typename Dtype>
-void caffe_gpu_rng_uniform(const int n, const Dtype a, const Dtype b, Dtype* r);
+void caffe_gpu_rng_uniform(const size_t n, const Dtype a, const Dtype b, Dtype* r);
 
 template <typename Dtype>
-void caffe_gpu_rng_gaussian(const int n, const Dtype mu, const Dtype sigma,
+void caffe_gpu_rng_gaussian(const size_t n, const Dtype mu, const Dtype sigma,
                             Dtype* r);
 
 template <typename Dtype>
-void caffe_gpu_rng_bernoulli(const int n, const Dtype p, int* r);
+void caffe_gpu_rng_bernoulli(const size_t n, const Dtype p, int* r);
 
 template <typename Dtype>
-void caffe_gpu_dot(const int n, const Dtype* x, const Dtype* y, Dtype* out);
+void caffe_gpu_dot(const size_t n, const Dtype* x, const Dtype* y, Dtype* out);
 
 template <typename Dtype>
-uint32_t caffe_gpu_hamming_distance(const int n, const Dtype* x,
+uint32_t caffe_gpu_hamming_distance(const size_t n, const Dtype* x,
                                     const Dtype* y);
 
 template <typename Dtype>
-void caffe_gpu_asum(const int n, const Dtype* x, Dtype* y);
+void caffe_gpu_asum(const size_t n, const Dtype* x, Dtype* y);
 
 template<typename Dtype>
-void caffe_gpu_sign(const int n, const Dtype* x, Dtype* y);
+void caffe_gpu_sign(const size_t n, const Dtype* x, Dtype* y);
 
 template<typename Dtype>
-void caffe_gpu_sgnbit(const int n, const Dtype* x, Dtype* y);
+void caffe_gpu_sgnbit(const size_t n, const Dtype* x, Dtype* y);
 
 template <typename Dtype>
-void caffe_gpu_fabs(const int n, const Dtype* x, Dtype* y);
+void caffe_gpu_fabs(const size_t n, const Dtype* x, Dtype* y);
 
 template <typename Dtype>
-void caffe_gpu_scale(const int n, const Dtype alpha, const Dtype *x, Dtype* y);
+void caffe_gpu_scale(const size_t n, const Dtype alpha, const Dtype *x, Dtype* y);
 
 #define DEFINE_AND_INSTANTIATE_GPU_UNARY_FUNC(name, operation) \
 template<typename Dtype> \
-__global__ void name##_kernel(const int n, const Dtype* x, Dtype* y) { \
+__global__ void name##_kernel(const size_t n, const Dtype* x, Dtype* y) { \
   CUDA_KERNEL_LOOP(index, n) { \
     operation; \
   } \
 } \
 template <> \
-void caffe_gpu_##name<float>(const int n, const float* x, float* y) { \
+void caffe_gpu_##name<float>(const size_t n, const float* x, float* y) { \
   /* NOLINT_NEXT_LINE(whitespace/operators) */ \
   name##_kernel<float><<<CAFFE_GET_BLOCKS(n), CAFFE_CUDA_NUM_THREADS>>>( \
       n, x, y); \
 } \
 template <> \
-void caffe_gpu_##name<double>(const int n, const double* x, double* y) { \
+void caffe_gpu_##name<double>(const size_t n, const double* x, double* y) { \
   /* NOLINT_NEXT_LINE(whitespace/operators) */ \
   name##_kernel<double><<<CAFFE_GET_BLOCKS(n), CAFFE_CUDA_NUM_THREADS>>>( \
       n, x, y); \
